+++ src/hapi/pipelines/database/food_security.py
@@ -104,10 +104,11 @@ class FoodSecurity(BaseUploader):
                                 reference_period_start=time_period_start,
                                 reference_period_end=time_period_end,
                                 population_in_phase=population_in_phase,
-                                population_fraction_in_phase=population_in_phase
-                                / population_total
-                                if population_in_phase
-                                else 0.0,
+                                population_fraction_in_phase=(
+                                    population_in_phase / population_total
+                                    if population_in_phase
+                                    else 0.0
+                                ),
                                 # TODO: For v2+, add to scraper (HAPI-199)
                                 source_data="not yet implemented",
                             )